depot initial

+++ database/migrations/2025_03_12_104722_create_notes_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreign('etudiant_id')->references('id')->on('users')->onDelete('cascade');
             $table->unsignedBigInteger('academic_year_id')->nullable();
             $table->foreign('academic_year_id')->references('id')->on('academic_years')->onDelete('cascade');
-            $table->unsignedInteger('type_evaluation_id')->nullable();
+            $table->unsignedBigInteger('type_evaluation_id')->nullable();
             $table->foreign('type_evaluation_id')->references('id')->on('type_evaluations')->onDelete('cascade');
             $table->decimal('note', 5, 2);
             $table->string('observation')->nullable();
